+++ packages/deepgram/src/deepgram-transcription-model.ts
@@ -10,7 +10,7 @@ import {
   parseProviderOptions,
   postToApi,
 } from '@ai-sdk/provider-utils';
-import { z } from 'zod/v4';
+import * as z from 'zod/v4';
 import { DeepgramConfig } from './deepgram-config';
 import { deepgramFailedResponseHandler } from './deepgram-error';
 import { DeepgramTranscriptionModelId } from './deepgram-transcription-options';