+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ### Patch Changes
 
+- Updated dependencies [95f65c2]
+- Updated dependencies [95f65c2]
+  - ai@6.0.0-beta.31
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.30
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - Updated dependencies [dee8b05]
   - ai@6.0.0-beta.29